refactor(mcp): split McpServerService into focused modules
Split the 3087-line monolith into 8 focused modules behind a thin facade.
Extracted: shared types/constants, tier authorization, audit persistence,
session store, renderer IPC bridge, terminal.waitUntilIdle, MCP session
server, and HTTP lifecycle. Facade preserves all public API, singleton
export, and IPC channel stability.

- shared.ts: types, tier allowlists, constants, pure helpers (~585 lines)
- tierAuth.ts: bearer token resolution, tool exposure, tier gating (~173 lines)
- auditLog.ts: audit ring buffer, classification, debounced persistence (~184 lines)
- sessionStore.ts: session/idle-timer Map ownership for both transports (~101 lines)
- rendererBridge.ts: IPC bridge for manifest/dispatch requests (~227 lines)
- waitUntilIdle.ts: native terminal.waitUntilIdle tool (~180 lines)
- sessionServer.ts: MCP SDK handler registration + resources/prompts (~780 lines)
- httpLifecycle.ts: HTTP server start/stop/supervision/retry/routing (~596 lines)
- McpServerService.ts: facade reduced from 3087 to ~304 lines
